Revert "Add MenuViewController"
